UPD: chess game
