--- OOP/chess.cc
@@ -1,6 +1,0 @@
-#include <iostream>
-#include "chman.h"
-int main() {
-    Figure k;
-    k.game();
-}
