+++ OOP/chman.h
@@ -2,22 +2,31 @@
 #define CHESSMAN
 
 #include <iostream>
-#include <string>
-using std::string;
+using std::cout;
 
 class Figure {
-    private:
-        char pos[2];
-        int cur_player;
+	protected:
+		char pos[2];
+	public:
+		Figure() {}; // нужен для реализации мн. наследования
+		Figure(char * p) {
+			pos[0] = p[0];
+			pos[1] = p[1];
+		};
+		bool operator==(char * p) { return (pos[0] == p[0]) && (pos[1] == p[1]); };
+        Figure& operator=(char*);
+		virtual int attack(char * p) = 0;
+		virtual char isa() = 0;
+		void desk();
+		static bool deskout(char * p);
+        char* get_pos() {return pos;};
+};
+
+class Rook: public Figure{
     public:
-        Figure() : cur_player(1) {pos[0] = 'h'; pos[1] = '8';};
-        bool operator==(string p) { return (pos[0] == p[0]) && (pos[1] == p[1]); };
-        static bool deskout(string p);
-        int attack(string p);
+        Rook(char* p) : Figure(p){};
+        int attack(char*);
         char isa() {return 'R';};
-        void desk();
-        void game();
-        void makeMove(string p);
 };
 
 #endif
